+++ tests/e2e/improved_csi_idempotency.go
@@ -958,10 +958,3 @@ func extendVolumeWithServiceDown(serviceName string, namespace string, client cl
 		gomega.Expect(quotavalidationStatus).NotTo(gomega.BeFalse())
 	}
 }
-
-// stopHostD is a function for waitGroup to run stop hostd parallelly
-func stopHostD(ctx context.Context, addr string, wg *sync.WaitGroup) {
-	defer ginkgo.GinkgoRecover()
-	defer wg.Done()
-	stopHostDOnHost(ctx, addr)
-}
